Changelog for Version 1.3.53 Important Update:
We have made a significant update to the build.gradle file to enhance compatibility with multiple Java versions. This change ensures that the plugin now supports Java 8, 16, 17, and 21. The script dynamically detects the server's Java version and sets the appropriate toolchain for compilation, ensuring broader compatibility and stability across different server environments. This update is crucial for maintaining the flexibility and robustness of the plugin in diverse deployment scenarios.